input method change
Because input is meant to work with bash, it proves itself a bad way of input, if you want to enter a multiple line text to analize.
In that case, input() takes only one line, and as soon as it reaches its end it stops reading, leaving all furtherly folloving text aside.
Trying to bypass this limitation I decided to use sys.stdin.read instead of input().
Note: there is another option - sys.stdin.readlines, but it doesn't fit because it returns a list with inputed lines as its elements. Also, it adds \n to every element and calculates them as extra "n"s, which is unacceptable in our case. Meanwhile .read gives proper effect.
However, there is a disadventage of using sys.stdin - the neccessity of entering Ctrl-D every time at inputing a text.
Nevertheless, further development of SymbCalc will use sys.stdin.read, till I find a better replacement for it.